Parametrize and rename TeacherTimetablesCleaner
